pyproject.toml
[tool.poetry]
authors = ["Arpad Fulop <data.ombudsman@tutanota.com>"]
description = "Implementation of IncrementalDBSCAN clustering."
license = "BSD-3-Clause"
name = "incdbscan"
version = "0.1.0"
readme = "README.md"
homepage = "https://github.com/DataOmbudsman/incdbscan"
repository = "https://github.com/DataOmbudsman/incdbscan"
keywords = [
    "clustering",
    "incremental clustering"
]
classifiers = [
    "License :: OSI Approved :: BSD License",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

[tool.poetry.dependencies]
networkx = "^2.8.7"
numpy = "^1.23.3"
python = ">=3.9.0,<4.0"
scikit-learn = "^1.1.2"
sortedcontainers = "^2.4.0"
xxhash = "^3.0.0"
opentelemetry-api = "^1.18.0"
faiss-cpu = "^1.8.0"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
jupyterlab = "^4.0.6"
line-profiler = "^4.1.1"
matplotlib = "^3.7.3"
pandas = "^1.5.3"
pylint = "^3.0.1"
pytest = "^7.4.2"
requests = "^2.26.0"
tqdm = "^4.66.1"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = ["testutils"]
profile = "pycharm"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]